Move the circe conversion methods into the json object
